SU-600: Make Factura Electronica Mensual job multi-threaded
