+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/jobs/facturaelectronicamensual/FacturaElectronicaMensualPoster.java
@@ -1,0 +1,153 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.fineract.portfolio.loanaccount.jobs.facturaelectronicamensual;
+
+import java.math.BigDecimal;
+import java.time.LocalDate;
+import java.time.YearMonth;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.fineract.infrastructure.core.service.DateUtils;
+import org.apache.fineract.infrastructure.jobs.exception.JobExecutionException;
+import org.apache.fineract.portfolio.loanaccount.invoice.data.LoanDocumentData;
+import org.apache.fineract.portfolio.loanaccount.service.LoanWritePlatformService;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+@Slf4j
+@RequiredArgsConstructor
+@Setter
+public class FacturaElectronicaMensualPoster {
+
+    private List<LoanDocumentData> loanInvoiceDataList;
+    private final JdbcTemplate jdbcTemplate;
+    private final LoanWritePlatformService loanWritePlatformService;
+
+    public void processInvoice() throws JobExecutionException {
+        List<Throwable> errors = new ArrayList<>();
+        try {
+            final String invoiceTransactionIds = loanInvoiceDataList.stream().map(LoanDocumentData::getTransactionIds)
+                    .collect(Collectors.joining(","));
+            final List<LoanDocumentData> groupedLoanInvoices = groupByClientIdAndProductType(loanInvoiceDataList);
+            for (final LoanDocumentData groupedLoanInvoice : groupedLoanInvoices) {
+                groupedLoanInvoice.setDocumentType(LoanDocumentData.LoanDocumentType.INVOICE);
+                this.loanWritePlatformService.processAndSaveLoanDocument(groupedLoanInvoice);
+            }
+            if (StringUtils.isNotBlank(invoiceTransactionIds)) {
+                final String updateLoanTransactionSQL = "UPDATE m_loan_transaction SET is_invoiced_generated_by_job = TRUE WHERE id IN ("
+                        + invoiceTransactionIds + ")";
+                final int rows = this.jdbcTemplate.update(updateLoanTransactionSQL);
+                log.info("Updated {} loan transactions as invoiced generated by job", rows);
+            }
+        } catch (Exception e) {
+            log.error("Failed to run Factura Electronica Mensual", e);
+            errors.add(e);
+        }
+        if (!errors.isEmpty()) {
+            log.error("Failed to run Factura Electronica Mensual", errors.get(0));
+            throw new JobExecutionException(errors);
+        }
+    }
+
+    private List<LoanDocumentData> groupByClientIdAndProductType(final List<LoanDocumentData> loanDocumentDataList) {
+        final LocalDate businessLocalDate = DateUtils.getBusinessLocalDate();
+        final YearMonth yearMonth = YearMonth.from(businessLocalDate);
+        final LocalDate lastDayOfMonth = yearMonth.atEndOfMonth();
+        final LocalDate firstDayOfMonth = businessLocalDate.withDayOfMonth(1);
+        final LocalDate secondLastDayOfMonth = lastDayOfMonth.minusDays(1);
+        return loanDocumentDataList.stream()
+                .collect(Collectors.groupingBy(li1 -> Arrays.asList(li1.getClientId(), li1.getProductTypeName()),
+                        Collectors.collectingAndThen(Collectors.toList(), list -> {
+                            final BigDecimal interestPaid = list.stream().map(LoanDocumentData::getInterestPaid).reduce(BigDecimal.ZERO,
+                                    BigDecimal::add);
+                            final BigDecimal penaltyChargesPaid = list.stream().map(LoanDocumentData::getPenaltyChargesPaid)
+                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                            final BigDecimal mandatoryInsurancePaid = list.stream().map(LoanDocumentData::getMandatoryInsurancePaid)
+                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                            final BigDecimal voluntaryInsurancePaid = list.stream().map(LoanDocumentData::getVoluntaryInsurancePaid)
+                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                            final BigDecimal honorariosPaid = list.stream().map(LoanDocumentData::getHonorariosPaid).reduce(BigDecimal.ZERO,
+                                    BigDecimal::add);
+                            final BigDecimal totalPaid = list.stream().map(LoanDocumentData::getTotalPaid).reduce(BigDecimal.ZERO,
+                                    BigDecimal::add);
+                            final BigDecimal penaltyChargesVatPaid = list.stream().map(LoanDocumentData::getPenaltyChargesVatPaid)
+                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                            final BigDecimal mandatoryInsuranceVatPaid = list.stream().map(LoanDocumentData::getMandatoryInsuranceVatPaid)
+                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                            final BigDecimal voluntaryInsuranceVatPaid = list.stream().map(LoanDocumentData::getVoluntaryInsuranceVatPaid)
+                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                            final BigDecimal honorariosVatPaid = list.stream().map(LoanDocumentData::getHonorariosVatPaid)
+                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
+                            final Integer loansCount = list.size();
+                            final LoanDocumentData loanDocumentData = list.get(0);
+                            return LoanDocumentData.builder().clientIdNumber(loanDocumentData.getClientIdNumber())
+                                    .productTypeParamId(loanDocumentData.getProductTypeParamId())
+                                    .billingPrefix(loanDocumentData.getBillingPrefix())
+                                    .billingResolutionNumber(loanDocumentData.getBillingResolutionNumber())
+                                    .rangeStartNumber(loanDocumentData.getRangeStartNumber())
+                                    .rangeEndNumber(loanDocumentData.getRangeEndNumber())
+                                    .lastInvoiceNumber(loanDocumentData.getLastInvoiceNumber())
+                                    .lastCreditNoteNumber(loanDocumentData.getLastCreditNoteNumber())
+                                    .lastDebitNoteNumber(loanDocumentData.getLastDebitNoteNumber())
+                                    .technicalKey(loanDocumentData.getTechnicalKey()).nota(loanDocumentData.getNota())
+                                    .nota(loanDocumentData.getNota()).clientDisplayName(loanDocumentData.getClientDisplayName())
+                                    .clientLastName(loanDocumentData.getClientLastName())
+                                    .clientLegalForm(loanDocumentData.getClientLegalForm()).clientId(loanDocumentData.getClientId())
+                                    .clientEmailAddress(loanDocumentData.getClientEmailAddress()).loanId(loanDocumentData.getLoanId())
+                                    .productTypeId(loanDocumentData.getProductTypeId())
+                                    .productTypeName(loanDocumentData.getProductTypeName())
+                                    .overdueSinceDate(loanDocumentData.getOverdueSinceDate())
+                                    .daysInArrears(loanDocumentData.getDaysInArrears()).interestPaid(interestPaid)
+                                    .penaltyChargesPaid(penaltyChargesPaid).mandatoryInsurancePaid(mandatoryInsurancePaid)
+                                    .voluntaryInsurancePaid(voluntaryInsurancePaid).honorariosPaid(honorariosPaid)
+                                    .penaltyChargesVatPaid(penaltyChargesVatPaid).mandatoryInsuranceVatPaid(mandatoryInsuranceVatPaid)
+                                    .voluntaryInsuranceVatPaid(voluntaryInsuranceVatPaid).honorariosVatPaid(honorariosVatPaid)
+                                    .totalPaid(totalPaid).loansCount(loansCount).firstDayOfMonth(firstDayOfMonth)
+                                    .secondLastDayOfMonth(secondLastDayOfMonth).lastDayOfMonth(lastDayOfMonth)
+                                    .loanProductName(loanDocumentData.getLoanProductName()).companyNIT(loanDocumentData.getCompanyNIT())
+                                    .companyDocType(loanDocumentData.getCompanyDocType())
+                                    .companyDeptCode(loanDocumentData.getCompanyDeptCode())
+                                    .companyDeptName(loanDocumentData.getCompanyDeptName())
+                                    .companyCityCode(loanDocumentData.getCompanyCityCode())
+                                    .companyCityName(loanDocumentData.getCompanyCityName())
+                                    .companyAddress(loanDocumentData.getCompanyAddress())
+                                    .companyTelephone(loanDocumentData.getCompanyTelephone())
+                                    .clientCedula(loanDocumentData.getClientCedula()).clientAddress(loanDocumentData.getClientAddress())
+                                    .clientCityCode(loanDocumentData.getClientCityCode())
+                                    .clientCityName(loanDocumentData.getClientCityName())
+                                    .clientFirstName(loanDocumentData.getClientFirstName())
+                                    .clientMiddleName(loanDocumentData.getClientMiddleName())
+                                    .clientTelephone(loanDocumentData.getClientTelephone())
+                                    .collectionHouseName(loanDocumentData.getCollectionHouseName())
+                                    .collectionHouseNit(loanDocumentData.getCollectionHouseNit())
+                                    .voluntaryInsuranceCode(loanDocumentData.getVoluntaryInsuranceCode())
+                                    .voluntaryInsuranceName(loanDocumentData.getVoluntaryInsuranceName())
+                                    .mandatoryInsuranceCode(loanDocumentData.getMandatoryInsuranceCode())
+                                    .mandatoryInsuranceName(loanDocumentData.getMandatoryInsuranceName())
+                                    .transactionIds(loanDocumentData.getTransactionIds()).build();
+                        })))
+                .values().stream().toList();
+
+    }
+}
