+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/jobs/facturaelectronicamensual/FacturaElectronicaMensualPosterTask.java
@@ -1,0 +1,49 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.fineract.portfolio.loanaccount.jobs.facturaelectronicamensual;
+
+import java.util.List;
+import java.util.concurrent.Callable;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.fineract.infrastructure.core.domain.FineractContext;
+import org.apache.fineract.infrastructure.core.service.ThreadLocalContextUtil;
+import org.apache.fineract.infrastructure.jobs.exception.JobExecutionException;
+import org.apache.fineract.portfolio.loanaccount.invoice.data.LoanDocumentData;
+
+@Slf4j
+@RequiredArgsConstructor
+public class FacturaElectronicaMensualPosterTask implements Callable<Void> {
+
+    private final FacturaElectronicaMensualPoster facturaElectronicaMensualPoster;
+    @Setter
+    private FineractContext context;
+
+    @Override
+    public Void call() throws JobExecutionException {
+        ThreadLocalContextUtil.init(context);
+        facturaElectronicaMensualPoster.processInvoice();
+        return null;
+    }
+
+    public void setLoanInvoiceDataList(List<LoanDocumentData> loanInvoiceDataList) {
+        this.facturaElectronicaMensualPoster.setLoanInvoiceDataList(loanInvoiceDataList);
+    }
+}
