+++ fineract-provider/src/main/resources/db/changelog/tenant/changelog-tenant.xml
@@ -351,4 +351,5 @@
     <include file="parts/2025013111428000000_SU_601_modify_columns.xml" relativeToChangelogFile="true"/>
     <include file="parts/2025021308508000000_SU_606_add_job_params_for_archivo_cartera_job.xml" relativeToChangelogFile="true"/>
     <include file="parts/2025130211009140_SU_605_add_collection_house_to_transaction.xml" relativeToChangelogFile="true"/>
+    <include file="parts/2025022412178000000_SU_600_add_job_params_for_factura_electronica_job.xml" relativeToChangelogFile="true"/>
 </databaseChangeLog>
